libs/digital-agency/agents/07_customer_support/community_manager/agent.py
"""
Community Manager Agent

Manages user communities, forums, and social engagement channels.
"""

from typing import Dict, List, Any, Optional
import yaml
from pathlib import Path


class CommunityManagerAgent:
    """
    Agent responsible for community management and engagement.

    Capabilities:
    - Manage community forums
    - Moderate discussions
    - Foster engagement
    - Organize events
    - Track community metrics
    """

    def __init__(self, config_path: Optional[str] = None):
        """Initialize the Community Manager Agent."""
        self.config = self._load_config(config_path)
        self.name = "Community Manager Agent"
        self.role = "community_manager"
        self.active_discussions = []

    def _load_config(self, config_path: Optional[str] = None) -> Dict[str, Any]:
        """Load agent configuration from YAML file."""
        if config_path is None:
            config_path = Path(__file__).parent / "config.yaml"
        try:
            with open(config_path, 'r') as f:
                return yaml.safe_load(f)
        except FileNotFoundError:
            return self._get_default_config()

    def _get_default_config(self) -> Dict[str, Any]:
        """Return default configuration."""
        return {
            'agent_name': 'Community Manager Agent',
            'model': 'gpt-4',
            'temperature': 0.6,
            'capabilities': ['community_management', 'moderation', 'engagement']
        }

    async def moderate_content(self, content: str) -> Dict[str, Any]:
        """Moderate community content."""
        return {'content_id': 'generated', 'approved': True, 'reason': ''}

    async def create_announcement(self, message: str, channels: List[str]) -> Dict[str, Any]:
        """Create community announcement."""
        return {'announcement_id': 'ANN-001', 'channels': channels, 'status': 'posted'}

    async def organize_event(self, event_data: Dict[str, Any]) -> Dict[str, Any]:
        """Organize community event."""
        return {'event_id': 'EVT-001', 'name': event_data.get('name'), 'status': 'scheduled'}

    def get_status(self) -> Dict[str, Any]:
        """Get current agent status."""
        return {
            'agent': self.name,
            'active_discussions': len(self.active_discussions),
            'status': 'active'
        }
