libs/digital-agency/pyproject.toml
[project]
name = "digital-agency"
version = "1.0.0"
description = "Digital Agency Automation Platform - AI-powered agency automation system"
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"

dependencies = [
    # Web Framework
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "pydantic==2.5.3",
    "pydantic-settings==2.1.0",

    # Database
    "sqlalchemy==2.0.25",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",

    # AI & API Clients
    "anthropic==0.18.1",
    "httpx==0.27.2",  # Updated to fix pytest-httpx compatibility
    "aiohttp==3.9.1",

    # Configuration & Environment
    "python-dotenv==1.0.0",
    "pyyaml==6.0.1",

    # Task Queue
    "redis==4.6.0",
    "celery==5.3.6",
    "flower==2.0.1",

    # Caching & State Management
    "redis-om==0.3.2",
    "hiredis==2.3.2",

    # Data Validation & Transformation
    "marshmallow==3.20.1",
    "jsonschema==4.20.0",
    "python-multipart==0.0.6",

    # Date & Time
    "python-dateutil==2.8.2",
    "pytz==2023.3",

    # HTTP & API Utilities
    "requests==2.31.0",
    "urllib3==2.1.0",

    # Authentication & Security
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "cryptography==42.0.0",
    "bcrypt==4.1.2",

    # Email
    "sendgrid==6.11.0",
    "jinja2==3.1.3",

    # Monitoring & Logging
    "sentry-sdk[fastapi]==1.39.2",
    "python-json-logger==2.0.7",
    "loguru==0.7.2",

    # API Integrations
    "google-api-python-client==2.116.0",
    "google-auth==2.27.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",

    # HubSpot
    "hubspot-api-client>=9.0.0,<13.0.0",

    # Slack
    "slack-sdk==3.26.2",

    # Social Media
    "tweepy==4.14.0",
    "facebook-sdk==3.1.0",

    # Data Processing
    "pandas==2.1.4",
    "numpy==1.26.3",
    "openpyxl==3.1.2",

    # File Handling
    "aiofiles==23.2.1",
    "python-magic==0.4.27",
    "pillow==10.2.0",

    # Utilities
    "tenacity==8.2.3",
    "python-slugify==8.0.1",
    "shortuuid==1.0.11",

    # Development
    "ipython==8.20.0",
    "ipdb==0.13.13",

    # Documentation
    "mkdocs==1.5.3",
    "mkdocs-material==9.5.5",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-httpx==0.30.0",  # Now compatible with httpx==0.27.2
    "pytest-watch==4.2.0",
    "faker==22.2.0",
    "factory-boy==3.3.0",

    # Code Quality
    "black==23.12.1",
    "flake8==7.0.0",
    "mypy==1.8.0",
    "isort==5.13.2",
    "pylint==3.0.3",
    "bandit==1.7.5",
    "safety>=3.0.0",

    # YAML Validation
    "yamllint==1.33.0",

    # Documentation
    "sphinx==7.2.6",
    "sphinx-rtd-theme==2.0.0",
    "pydoc-markdown==4.8.2",

    # Development Tools
    "jupyter==1.0.0",
    "notebook==7.0.6",

    # File Watching
    "watchdog==3.0.0",

    # Type Stubs
    "types-requests==2.31.0.10",
    "types-redis==4.6.0.11",
    "types-PyYAML==6.0.12.12",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "config*", "core*", "shared*", "workflows*"]
exclude = ["tests*", "docs*", "scripts*", "deployment*", "monitoring*", "infrastructure*"]

[tool.black]
line-length = 127
target-version = ['py311', 'py312']
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | env
  | __pycache__
  | \.pytest_cache
  | \.mypy_cache
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 127
skip = [".venv", "venv", "env", "__pycache__"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "env",
    "tests",
]

[tool.pylint.messages_control]
max-line-length = 127
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
]

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/.venv/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
